Tidy some functions related to language code handling
